pyproject.toml
[tool.poetry]
name = "event-tpc"
version = "0.1.0"
description = ""
authors = ["SharpKoi <sharpkoi1222@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
polars = "^1.21.0"
torch = "^2.6.0"
matplotlib = "^3.10.0"
scikit-learn = "^1.6.1"
tqdm = "^4.67.1"
torchmetrics = "^1.6.1"
wandb = "^0.19.5"
mlflow = "^2.20.1"
ruamel-yaml = "^0.18.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
